backend/requirements.txt
Flask==3.0.3
Pillow==10.4.0
python-dotenv==1.0.1
ultralytics==8.3.4
